What the fuck is going on
See line 26 of slide.c: I use VLAs to store the onscreen coordinates of
gridlines. Now see line 29 of drawing.h, and notice how I don't use the
variable i anywhere in that function. Also note that, as is, the '+'
symbols appear as intended. However, if you delete line 29 in drawing.h,
they don't! Furthermore, if you go back to slide.c and malloc the line
drawing arrays rather than setting them as VLAs, line 29 in drawing.h is
not necessary for the program to work.